Ajout de smewzzzzzzzzzzz

+++ rflpc-master/samples/test_ethernet/wait.h
@@ -1,1 +1,1 @@
-int wait(int sec);
+int wait(double sec);
